model_service/requirements.txt
numpy~=2.3.5
joblib~=1.4.2
lightgbm~=4.6.0
scikit-learn~=1.7.2
Django~=5.1.2
pandas~=2.3.3
requests~=2.32.5
matplotlib~=3.10.7
seaborn~=0.13.2
djangorestframework~=3.15.2
